memory cache implementation.
This  cache uses shmem and performs well when running concurrently seeds in single test.
The reason for shmem is because concurrently is a multiprocess and not multithread.
This cacher does not handle seed dependency as it assumes that code does not change during the test.
This issue must be handled for DEVELOPER_MODE in systemtests where code actually do change during test