bugfix: avoid long filenames that can exceed linux maximum 255
Linux filenames are maximized to 255 characters, when a lot of flags are added and filename is large, cache filenames
could brake this limit thus avoiding seedcache from working.
The reason for encoding everything in the name was to allow to do decoding and display seed cache parameters and names
(display cache command)
In order to avoid the issue we will "encode" the name in md5 (and hope for no colissions:)) and pack the flags into manifest file with the dependencies records, so that we can decode them.
MD5 will be performed on all parameters.