bugfix: PYTHONPATH order was overriden by the seedcreator
Bug was introduced in change: 4d2b56306ec2c40dec524f9aaaf76bcb61ba00aa in order to
supposed scripts without namespace join. This had altered the behaviour of the seedcreator
that relied on the origianl pythonpath, and therfore took upseto.manifest from the first matched module according to the pythonpath,
set by the test. the change '4d2b56306ec2c40dec524f9aaaf76bcb61ba00aa' altered this behaviour so that the actual seed would be the first in the path.
This broke some tests in projects that relied on the original behaviour.